tester comments ++ aswell as refractoring
